requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
anytree==2.8.0
apkutils==0.11.0
attrs==21.2.0
beautifulsoup4==4.10.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colored==1.4.3
cryptography==36.0.0
decorator==5.1.0
Deprecated==1.2.13
facebook-wda==1.4.6
idna==3.3
imageio==2.13.3
iniconfig==1.1.1
Jinja2==3.0.3
logzero==1.7.0
lxml==4.6.4
MarkupSafe==2.0.1
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyelftools==0.27
pyftype==1.2.3
Pygments==2.10.0
pyparsing==3.0.6
pytest==6.2.5
requests==2.26.0
retry==0.9.2
simple-tornado==0.2.2
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
tidevice==0.5.6
toml==0.10.2
tornado==6.1
urllib3==1.26.7
wrapt==1.13.3
xmltodict==0.12.0
